image_capture/requirements.txt
opencv-python
minio
psycopg2-binary
numpy
torch
torchvision
transformers
